# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Delivery Time App API docs",
    "description": "Delivery Time App API docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/{userId}/admin/{boolFlag}": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "Toggle ADMIN",
        "description": "By calling this endpoint, you can toggle admin privileges on account\n",
        "operationId": "toggleAdmin",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "pass an id of user",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "boolFlag",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated."
          },
          "400": {
            "description": "Error."
          }
        }
      }
    },
    "/authorize": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Authorize - get JWT token",
        "description": "By calling this endpoint, you can login to get JWT token\n",
        "operationId": "authorize",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "pass account username",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "pass account password",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JTW Token"
          },
          "400": {
            "description": "Error."
          }
        }
      }
    },
...
  },
...
}